# packages/contract/generated/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Clone Instagram API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Même origine ; l'hôte dépend de l'environnement (local ou démo, ADR-009)."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/me/avatar": {
      "delete": {
        "operationId": "removeAvatar",
        "tags": [
          "identity"
        ],
        "summary": "Retirer ma photo de profil",
        "description": "Retire la photo de profil ; l'ancien média est détaché puis purgé (ADR-008). Sans photo, ne change rien.",
        "responses": {
          "200": {
            "description": "Profil sans photo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Profil de l'utilisateur authentifié.",
                  "additionalProperties": false,
                  "required": [
                    "id",
                    "username",
                    "fullName",
                    "bio",
                    "birthDate",
                    "isPrivate",
                    "status",
                    "followerCount",
                    "followingCount",
                    "postCount",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "Identifiant Supabase Auth."
                    },
                    "username": {
                      "type": "string",
                      "description": "1 à 30 caractères `[a-z0-9._]`, en minuscules (ADR-007).",
                      "minLength": 1,
                      "maxLength": 30,
                      "pattern": "^[a-z0-9._]{1,30}$",
                      "examples": [
                        "killian.b"
                      ]
                    },
                    "fullName": {
                      "type": "string",
                      "description": "Nom affiché, 1 à 30 caractères, au moins un caractère autre qu'un espace.",
                      "minLength": 1,
                      "maxLength": 30,
                      "pattern": "\\S",
                      "examples": [
                        "Killian Boularand"
                      ]
                    },
                    "bio": {
                      "type": "string",
                      "description": "150 caractères au plus ; chaîne vide si aucune bio.",
                      "maxLength": 150
                    },
                    "birthDate": {
                      "type": "string",
                      "format": "date",
                      "description": "Renvoyée à son seul propriétaire."
                    },
                    "isPrivate": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "suspended",
                        "banned"
                      ]
                    },
                    "followerCount": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "followingCount": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "postCount": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "avatar": {
                      "description": "Photo de profil ; absente si aucune photo.",
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "thumb",
                        "medium",
                        "large"
                      ],
                      "properties": {
                        "thumb": {
                          "type": "string",
                          "format": "uri",
                          "description": "150 px de large."
                        },
                        "medium": {
                          "type": "string",
                          "format": "uri",
                          "description": "640 px de large."
                        },
                        "large": {
                          "type": "string",
                          "format": "uri",
                          "description": "1080 px de large."
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide (`unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Onboarding non terminé (`profile_not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur inattendue (`internal_error`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/usernames/{username}/availability": {
      "get": {
        "operationId": "getUsernameAvailability",
        "tags": [
          "identity"
        ],
        "summary": "Disponibilité d'un username",
        "description": "Vérifiée en direct pendant l'onboarding. Si le username est pris, jusqu'à 3 usernames libres dérivés du candidat sont suggérés. Un username au mauvais format est refusé en `400`. Le username de l'appelant lui-même est disponible pour lui.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "1 à 30 caractères `[a-z0-9._]`, en minuscules (ADR-007).",
              "minLength": 1,
              "maxLength": 30,
              "pattern": "^[a-z0-9._]{1,30}$",
              "examples": [
                "killian.b"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disponibilité du username.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "username",
                    "available",
                    "suggestions"
                  ],
                  "properties": {
                    "username": {
                      "type": "string",
                      "description": "1 à 30 caractères `[a-z0-9._]`, en minuscules (ADR-007).",
                      "minLength": 1,
                      "maxLength": 30,
                      "pattern": "^[a-z0-9._]{1,30}$",
                      "examples": [
                        "killian.b"
                      ]
                    },
                    "available": {
                      "type": "boolean"
                    },
                    "suggestions": {
                      "type": "array",
                      "description": "Vide si le username est disponible ; sinon jusqu'à 3 usernames libres.",
                      "maxItems": 3,
                      "items": {
                        "type": "string",
                        "description": "1 à 30 caractères `[a-z0-9._]`, en minuscules (ADR-007).",
                        "minLength": 1,
                        "maxLength": 30,
                        "pattern": "^[a-z0-9._]{1,30}$",
                        "examples": [
                          "killian.b"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrée non conforme au schéma (`validation_failed`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide (`unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur inattendue (`internal_error`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token Supabase Auth."
      }
    },
...
  }
}